Making a new "iohandler" branch since I have completely replaced the old InputReader stack with a new IOHandler stack.
Details to come when I check the changes in. We will need all of the major systems (MacOSX, Linux, FreeBSD, and Windows) to check this branch out and try it out and make sure the test suite runs ok. 

Once we get the ok from all systems, we can merge this back into top of tree.